Move log-manager.js
