+++ src/grammar-support/log-manager.js
